requirements.txt
streamlit-nightly==1.36.0
streamlit-authenticator
openai
requests